+++ lib/src/utils/date_util.dart
@@ -1,5 +1,0 @@
-
-bool isDateExpired(int startDate, int duration) {
-  final today = DateTime.now().millisecondsSinceEpoch;
-  return today - startDate > duration;
-}